+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Removed typo/deprecated enum value MusicErrorType.Timesignature.
 - Removed deprecated types StaffTabOrGroup (-> StaffTarget), StaffTabOrGroups (-> StaffTargets).
 - Removed deprecated enum Fermata and addFermata (-> addAnnotation) and addFermataTo (-> addAnnotationTo).
+- Removed deprecated enum value AnnotationKind.measureEndFermata (-> { anchor: "rightBarLine" }).
 
 ## [6.4.1] - 2026-03-22
 ### Fixed